# Conflicts:
#	TESTERS.md
#	app/src/main/resources/version.txt

+++ app/src/main/java/elite/intel/gameapi/journal/events/CarrierJumpEvent.java
@@ -147,7 +147,7 @@ public class CarrierJumpEvent extends BaseEvent {
         return starSystem;
     }
 
-    public double getSystemAddress() {
+    public long getSystemAddress() {
         return systemAddress;
     }
